Add first confirm for initial failing test
The intent is to add a random string to the message which makes it
unique. To confirm the message is logged, need to read the log file and
search for the message containing the same random string.

This requires some utility functionality to read the log file and search
for the string. Currently, this is not implemented, hence the failing
test. Also, the generation of a range string to add to the message is
also not implemented yet.